discord_news_bot/src/lambda/requirements.txt
requests
discord.py
atproto
pynacl
boto3
feedparser